export/pyproject.toml
[project]
name = "hlab-export"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["torch==2.3.*","torchvision==0.18.*","pyyaml>=6.0"]
